requirements.txt
fastapi
uvicorn
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
blinker==1.9.0
build==1.2.2.post1
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
cleo==2.1.0
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
crashtest==0.4.1
dataclasses-json==0.6.7
Deprecated==1.2.18
distlib==0.3.9
distro==1.9.0
dulwich==0.22.8
durationpy==0.9
fastapi==0.115.12
fastjsonschema==2.21.1
filelock==3.18.0
filetype==1.2.0
findpython==0.6.3
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.3.2
google-ai-generativelanguage
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.10.0
google-generativeai
googleapis-common-protos==1.69.2
greenlet==3.1.1
groq==0.22.0
grpcio==1.72.0rc1
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.6
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
keyring==25.6.0
kubernetes==32.0.1
langchain==0.3.23
langchain-chroma==0.2.2
langchain-community==0.3.21
langchain-core==0.3.51
langchain-google-genai
langchain-groq==0.3.2
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.8
langsmith==0.3.27
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.2
mypy-extensions==1.0.0
networkx==3.4.2
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.21.0
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-asgi==0.52b1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-util-http==0.52b1
orjson==3.10.16
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pbs-installer==2025.3.17
pillow==11.1.0
pkginfo==1.12.1.2
platformdirs==4.3.7
poetry==2.1.2
poetry-core==2.1.2
posthog==3.23.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
Pygments==2.19.1
pyparsing==3.2.3
pypdf==5.4.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.0.2
setuptools==78.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
sympy==1.13.1
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.1
tomlkit==0.13.2
torch==2.6.0
tqdm==4.67.1
transformers==4.51.1
trove-classifiers==2025.3.19.19
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.30.0
watchfiles==1.0.5
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.19.0
zipp==3.21.0
zstandard==0.23.0
